agen-sme-eval-be/pyproject.toml
[project]
name = "agen-sme-eval-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.1",
    "asyncio-throttle==1.0.2",
    "autogen-core==0.7.4",
    "black==23.11.0",
    "fastapi==0.104.1",
    "flake8==6.1.0",
    "httpx==0.25.2",
    "litellm==1.51.0",
    "mypy==1.7.1",
    "pydantic>=2.10.0,<3.0.0",
    "pydantic-settings>=2.5.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "python-dotenv==1.0.0",
    "pyyaml==6.0.2",
    "structlog==23.2.0",
    "uvicorn[standard]==0.24.0",
]
